Version guards with merged HCR/OSR guards only when possible
Loop versioner has been versioning virtual guards with merged HCR or OSR
guards without taking the merged guards into account.

For example, consider a nonoverridden guard with a merged HCR guard. If
the receiver is invariant, then it would be versioned, and the merged
HCR guard would effectively be versioned along with it even if HCR guard
versioning is not supposed to be possible in the loop, i.e. even if the
hot loop will still have HCR invalidation points in it after versioning.
So if the protected method were redefined during execution of the
resulting hot loop, the loop would continue to run the old definition
until it exited.

With this commit, virtual guards with merged HCR guards will now be
versioned only when HCR guard versioning is possible, and those with
merged OSR guards will now be versioned only when OSR guard versioning
is possible.