AArch64: Handle unaligned immediate offset memory access case
On aarch64, the immediate offset to the memory access must be
the multiples of 16 when the size of access is 128-bit.
This commit changes commonLoadEvaluator and commonStoreEvaluator
to handle those cases manually.
In the longer term, MemoryReference class should take the size parameter
and take it into acount when populating.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>