Define ivLoadSeesUpdatedValue() based on blockIsAlwaysExecutedInLoop()
This method determines whether a given load of a versionable induction
variable sees the updated value or the iteration-initial value. This
determination will be necessary in order to properly version conditional
trees with respect to extrema, e.g. isVersionableIfWithMaxExpr().

The result is determined mostly by the set of prior blocks found by the
new implementation of blockIsAlwaysExecutedInLoop().