Stop trying to version bound checks w.r.t. multiply-updated IVs
This logic was added long ago in order to pick up on induction variables
identified by an earlier pass of global value propagation, specifically
in order to allow versioning in cases where there are multiple IV
updates within the loop.

But we do not version in such cases. When we generate the versioning
test in buildBoundCheckComparisonsTree() for a non-invariant index,
_storeTrees is consulted unconditionally to find the IV update store
node. Any attempt to version w.r.t. an IV that has multiple updates
within the loop would cause a crash.

Furthermore, the possibility of multiple IV updates complicates the
determination of old/unchanged vs. new/updated IV value at a load. And
worse, it turns the old vs. new question into a false dichotomy, since
there is a possibility of multiple IV updates on a single path through
the loop, which would allow some loads to produce neither value.

Loop strider is the only other optimization that adds to the list of IVs
consulted here. Loop strider typically creates address or 64-bit integer
IVs, which are already ignored by versioner. (The address IVs aren't
even added to the list in the first place.)

This removal does not interfere with versioning w.r.t. any IV that has
just a single update within the loop, since versioner can identify those
on its own.