Replace boundCheckUsesUnchangedValue() with ivLoadSeesUpdatedValue()
The implementation of boundCheckUsesUnchangedValue() was hard to follow,
and it was also buggy. Following are a couple of example loops in which
a bound check using the unchanged IV value would have previously been
incorrectly identified as using the updated IV value instead.

Example 0:

    BBStart <loopHeader>
    treetop
      iadd
        iload i
        iload i (*)
    istore i
      isub
        iload i
        iconst -1
    BNDCHK
      ...
      ==>iload i (*)
    ...
    BBEnd </loopHeader>

Example 1:

    BBStart <loopHeader>
    istore tmp
      iload i
    ...
    BBEnd </loopHeader>

    ...

    BBStart <loopTestBlock>
    istore i
      isub
        iload i
        iconst -1
    BNDCHK
      ...
      iload tmp
    ...
    BBEnd </loopTestBlock>

These are just the problems that I found relatively easily; there might
have been even more bugs hiding in the old (now deleted) logic.