Pass task port for core dump using IPC
In MacOS 13, we can no longer pass the task port to a child process
through setting the bootstrap port to the same port name as the task
port and having the child inherit that. Instead we must pass a normal
port to the child process which will be creating the core dump in order
to set up messaging. Then we can use that to pass the task port so the
child has access to the parent.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>