Use LDE to load single precision floating point values
LE instruction on IBM Z loads single precision floating values into the
FPR register. In certain scenarios this operation may perform poorly
espcecially in case of Write after Read or Write after Write scenario
where hardware may not be able to break the dependency chain. In such
scenarios, using LDE instruction to load from memory would help. This
commit replaces the use of LE instruction with LDE.
Also while testing the changes we found out that for RXE type
instructions we were not handling long displcement correctly and
introducing a bug where it would store the scratch register to stack but
does not restore it back. It is fixed in this commit.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>