Disable CLassLookahead because it is possible for static final field value to be changes via reflection or the unsafe API.
1. Disable ClassLookahead. Similar to GVP, ClassLookahead is using
static analysis of static final field initialization code to provide
field information that is later used to make assumptions about the
field contents which might not be true if reflection or the unsafe
API is used to change static final field values. An environment
variable is introduced to allow ClassLookahead to be re-enable.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>