Avoid overflow calculating number of loop iterations
If the bounds on a loop induction variable differ by about 2^31 or more,
the calculation of the iteration count in LoopInfo::getNumIterations
might overflow.  Changed calculation of number of iterations to use
int64_t intermediate values and then check whether the result is less
than or equal to the maximum int32_t value.  If it is greater, a result
of zero is returned to indicate that the number of iterations is
unknown.