Check for right-hand operand of zero in lrem and irem
The constrainIrem and constrainLrem methods guard against a right-hand
operand of zero if the left-hand operand is a constant, but not if the
left-hand operand is a non-constant.  In the latter case, a constraint
in the range [1,-1] results.

This change adds an additional check for a right-hand operand equal to
zero if the left-hand operand is non-constant, avoiding creating an
erroneous constraint for the result.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>