Add _numAllocatedCompilationThreads field
Add the _numAllocatedCompilationThreads field to allow the compiler to
distinguish between the number of usable compilation threads vs the
number of allocated compilation threads. This, in essence, means that
_numUsableCompilationThreads can be used as a "soft limit", whereas
_numAllocatedCompilationThreads can be used as a "hard limit". By
default though, these two fields are set to be the same value.

A more tangible use case is when implementing Checkpoint/Restore
functionality; this new field allows a runtime to change the number of
compilation threads post restore.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>