Implement regex matching based on a location in a method
This change defines a new SimpleRegex::match method that will match a
regular expression against a location in a sequence of inlined method
invocations that are described by a TR_ByteCodeInfo.  The
TR_ByteCodeInfo is rendered as strings of the form

and

@<bc-offset>{#<inlined-method-sig>@<bc-offset>}*

where each <bc-offset> is the bytecode offset from the start of a
method, and each <method-sig> is the signature of the outermost method
or an inlined method invocation.  If either string matches the specified
regular expression, the match is successful.  That makes the outermost
method signature optional in the specified regular expression.