Use non-asserting query to print nodeCreatedByPRE
In nodePrintAllFlags(), chkNodeCreatedByPRE() is used to check
for the presence of the flag, but its current implementation  will
assert (on a debug build) whenever the flags are printed on a node
for which isLoadVarDirect() is false. nodePrintAllFlags() should use
the form of the query that simply returns false if the query is invalid
for that node. Other node flag methods have a chk<flag> query that does
not assert and an is<flag> query that does assert for the validity of
checking the requested flag. This change makes the node flags methods
for the nodeCreatedByPRE flag follow the same approach as other node
flag query methods so that chkNodeCreatedByPRE() does not assert while
isNodeCreatedByPRE() does.