Check receiver class initialized or not before checking overridden bits
`updateOverriddenFlag` is called by `updateCHTable` to update
overridden bits. `updateCHTable` is triggered either by
`jitHookClassLoad` or `jitHookClassPreinitialize`. For fixed classes,
we need to make sure the class is initialized so that the overridden
bits are up to date before calling `virtualMethodIsOverridden`.
Otherwise, `canFoldNonOverriddenGuard` could fold away non-overridden
guards when a method is already overridden but the JIT thinks otherwise
because the overridden bits has not been updated.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>