GC Reinitialize Sweep For Restore (CRIU)
Reinitialize and update the sweep sectioning data, sweep chuck size and
pool, to reflect the adjusted thread count for restore (for background,
see https://github.com/eclipse/omr/issues/6888). This requires the chunk
pool growth and utilization order to be changed, from LIFO to FIFO
style. The pool now grows at the tail and is consumed starting at the
head. This is necessary because the baseArray (first inited pool array)
is shared with compaction. Compaction looks to the used size of the this
baseArray to determine size it can use (getBackingStoreSize()). In a
traditional VM run, backing store size will never be 0 (as asserted by
compact). However,  if the pool continued to use LIFO style with CRIU,
it could result in unused baseArray and backing store size of 0.
Specifically, when a decrease in thread count results in increased chunk
size/less utilized chunks. FIFO will guarantee that the baseArray is
utilized first.

Signed-off-by: Salman Rana <salman.rana@ibm.com>