Add ArrayTest suite to compilertriltest
This suite contains tests for arraycmp evaluator.
One of the tests verifies that arraycmp evaluator works correctly
on arrays with the same data.
The other verifies that it properly returns the offset of the
mismatched element or expected value for greater than/less than cases
when the passed arrays are not equal.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>