Update documentation to reflect changes to copyright notices
There is no longer any need to update the year in the copyright notice
at the top of each modified file.

The copyright/license file header template is updated so that the
copyright notice in it will be consistent with the IBM copyright notices
in most files (once they are updated in a subsequent commit).