AArch64: Implement arrayset evaluator
This commit adds arrayset evaluator to aarch64 codegen.
- For variable length arrays, the main loop writes a 64-byte chunk
to the array in a single interation. It uses stp instructions which
store 32-byte data using two 128-bit vector registers.
If the length of the array is in between 16 and 96 bytes, then it writes
appropriate size of chunks from both the beginning and end of the array.
If the length of the array is less than 16 bytes, then the loop which writes
a single byte per iteration is executed.
- For constant length arrays, if the length is greater than or equal to
32 bytes, stp instructions with vector registers are used.
If the length is larger than certain threshold, then the loop which stores
256 bytes in a single iteration is executed.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>