Refine unsafe shadow to a more precise shadow in VP whenever possible
This improves aliasing and when loading a reference it opens the door
for determining a bound on the type of the referent based on the
declared type of the field (though that information is not yet available
from fabricated field symrefs).

VP no longer populates the set of unsafeArrayElementSymRefs(), since
whenever it could determine that the base object of an unsafe access is
an array, it should have had a specific array type for it. In that case
the unsafe shadow will now be replaced with the correct array shadow
instead.

Since nothing else populated unsafeArrayElementSymRefs(), all uses of it
are removed. For the moment the definition remains in order to avoid
breaking OpenJ9, which still has a single (harmless) use.