x86: fix mem-ref binary length estimation
If a memory reference is VFP-relative and an
address load instruction is required, estimating
the binary length requires the VFP register to be
assigned. On AMD64, the estimation logic was produced
by a conservative constant and did not inherit the
logic which applies to the vast majority of memory
references. By not calling the inherited logic,
VFP was not assigned.

Signed-off-by: Bradley Wood <bradley.wood@ibm.com>