Consider array size that might be outside range of valid values
The Value Propagation handlers for newarray and anewarray assume that if
a size constraint exists for the array length, and it's not definitely
negative and definitely does not exceed the maximum possible array size,
the operation must not throw an Exception or Error.  In such cases, they
set the AllocationCanBeRemoved flag for the node, which allows Dead
Trees Elimination to delete the node later if the array is not used.

However, even if the length is not definitely negative or not definitely
too great, does not mean that it's definitely non-negative nor definitely
not too great - it might have a range that includes "reasonable"
non-negative values, but also includes values that are negative or too
great.

Fixed this by only setting the AllocationCanBeRemoved flag if the lower
bound of array length is non-negative and the upper bound does not
exceed the maximum possible array size.