Add TR_StartPC relocation
This relocation is responsible for updating the startPC of the method
being compiled. For now it is intended to be used on power only. This
relocation cannot be accomplished with TR_AbsoluteMethodAddress
because, among other things, the relocation infrastructure currently
lacks support for loading address sequences from code binaries on power.

Signed-off-by: Christian Despres <despresc@ibm.com>