Have GLU consume budget only for successfully unrolled loops
Otherwise, it's possible for loops that can't actually be unrolled to
use up budget and prevent unrolling in other loops. The two overloads of
TR_LoopUnroller::unroll() both already indicate whether unrolling was
successful, but that result was ignored.

Also trace when GLU skips a loop for being too expensive to unroll.