Fix bug in updating displacement field in RSY type instruction
In https://github.com/eclipse/omr/pull/6893, encoding on RSY type
instruction was fixed to allow a immediate field to be used with RSY
type instructions. In encoding routine for that instruction,
miscalculation in updating location for DH2 field in instruction caused
incorrect immediate value being encoded in the instruction.

Fixes: https://github.com/eclipse-openj9/openj9/issues/16969

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>