Check const values to ensure correct pattern in selectSimplifier
...when simplifying a select of a const 0/1 and a boolean expression.
The simplification only works when the const is either 0 or 1, but the
existing code only checks for whether there is a const initially, and
then for whether that const's value is either 0 or something else. This
allows consts that are not either 0 or 1 to be used incorrectly in this
simplification.