Ensure DIVCHK and division/remainder simplifiers uphold contract
The divchkSimplifier was first simplifying its child node and then
checking whether the DIVCHK is still needed after having simplified its
child.  It did that by checking whether the simplification of its child
results in a node that is not the original child, or the simplified
child is not a division or remainder operation.

However, it might be that the simplified child is still a division or
remainder operation that needs to be checked for division by zero, but
is not the original child node.  That leads divchkSimplifier to replace
the DIVCHK with a treetop operation incorrectly.

This change introduces a _nodeToDivchk field in OMR::Simplifier that the
integer division and remainder simplifiers use to indicate whether the
result of attempting the simplification results in a node that would
still need to be checked for division by zero if the original node
needed to be checked for division by zero.

In most cases that handling of the handshake in the integer division and
remainder simplifiers happens in a new function that's called if the
divisor is a non-zero constant, permitSimplificationOfConstantDivisor.
That function is responsible for checking that the parent of the
operation is a DIVCHK, and if so, that performTransformation allows the
DIVCHK to be removed.

The divchkSimplifier, after simplifying its child division or remainder
operation, checks _nodeToDivchk as a first step in determining whether
the DIVCHK is still needed.  If it is NULL, the DIVCHK is removed; if it
is non-NULL, that node is made the subject of the DIVCHK.

Also, moved references to BINARY_IDENTITY_OP in the integer division
and remainder simplifiers so that it's only used in situations where the
divisor is already known to be a constant, so the handshake with the
divchkSimplifier will already have been handled by those simplifiers.
That avoids having BINARY_IDENTITY_OP deal with the handshake itself.

This change also fixes bugs in some simplifiers where division by zero
could result from attempting to fold the result of an operation where
both operands are constant.