Update trace message to reflect value of hasCommonedAncestor
OMR::Optimization::anchorChildren sets hasCommonedAncestor to the result
of the comparison of the node's reference count and 1.  However, the
traceMsg it was emitting claimed that it was always setting
hasCommonedAncestor to true, even if the result of the comparison was
false.  Corrected this to reflect that actual result of the comparison.