Update System.arraycopy to support null restricted value type arrays
Required by JEP 401: If null restricted value type is enabled,
we need to preform null check on the value being stored in order to
throw a `NullPointerException` if the array is null-restricted
and the value to write is null. If it is this case,
`System.arraycopy` cannot be transformed into arraycopy instructions.

Based on whether or not the source array and the destination array
are null restricted value type, and whether or not they are flattened,
VP decides whether or not transform `System.arraycopy`.

Fixes: eclipse-openj9/openj9#17196,eclipse-openj9/openj9#17586

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>