Prevent ldiv/i2l -> i2l/idiv simplification when idiv could overflow
Simplifier detects the following pattern and rewrites it:

    ldiv                                i2l
      i2l                                 idiv
        <dividend>          ==>             <dividend>
      i2l                                   <divisor>
        <divisor>

But this is correct only if the resulting idiv will never overflow. The
original tree always produces the true mathematical result.

To rule out overflow of idiv, simplifier now requires that at least one
of the operands be nonnegative, since signed division overflows only
when dividing the minimum signed value by -1.