Fixed shift calculations for shift right on Power
In some cases when attempting a shift right calculation, a rotate left
instruction is used instead on Power. For this to work, the shift value needs
to be convered to be (data width - shift amount).

eg. Shift right by 4 on a 32 bit value becomes a rotate left and mask by
(32-4=) 28.

However, this doesn't work when trying to shift right by 0. It gets converted
to a rotate left and mask by 32 (for 32 bit values) which is not a legal
encoding of the instruction.

This has now been updated to take the modulo by data width after the old
version of the conversion. This will catch cases of a shift right by 0 and
convert to a rotate left and mask by 0.

If it can be determined the the shift by 0 will not change the result in the
source register and the source can be cloobered, the register is simply passed
on and no instructions are generated. If the source can not be clobbered, a
register move is generated instead.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>