AArch64: Fix Vector Rotate Evaluator
`vrolEvalutor` on AArch64 uses a left shift and a right shift operations to
achieve a vector rotate operation.
However, if elements of the second source vector which describes rotate amount
are negative, it incorrectly calculates the shift amount applied fo
the left and right shift operations.
This commit fixes the evalutor to validate the rotate amount vector
before applying shift operations.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>