AArch64: Add a parameter for generating relocatable code to loadAddressConstant
The existing `loadAddressConstant` function on AArch64 determines
whether it should generate relocatable code by the answer returned
by `compileRelocatableCode()` query. However, it does not return
the correct answer under a remote compilation.
To support remote compilations, this commit adds an overloaded version
of `loadAddressConstant` with a boolean parameter that allows
the caller of the api to determine if the relocatable code is generated.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>