Nest percolate gc for aborted scavenge
Issue: The global gc stats after an aborted scavenge currently
reported incorrect reclaimed memory, and led to a situation
that the excessive gc is not triggered while it should be.

Reason: The excessive gc knows that it's possibly inside
a recursive call and should use the free space before
the outermost gc call to calculate. However, the first
aborted scavenge completes and the second scavenge is
treated as the outermost gc, which leads to mistakenly
recorded "before free space".

We now call the global gc recursively right after the
aborted scavenge instead of the next scavenge. In this
way it could correctly calculate the reclaimed memory.
The formula is now:
reclaimed memory = free space after global - free
space before the aborted scavenge