AArch64: Fix helper trampoline lookup fo ARM64Trg1SymImmInstruction
OMR #7077 introduced the `callUsesHelperImplementation()` query to add
a `TR_Helper` relocation for methods that are actually helper calls.
That PR also changes the way the binary encoder finds necessary trampolines.
Before OMR #7077, `methodTrampolineLookup()` was used for call targets
whose method object returns false for the `isHelper()` query. The behavior
was changed to call `findHelperTrampoline()` instead of `methodTrampolineLookup()`
if the `callUsesHelperImplementation()` query returns true.

The problem is that the symbol reference for a certain method (which is
actually a helper call) does not have a valid helper index as the reference
number, thus causing `findHelperTrampoline` to return an incorrect address.

This commit fixes the binary encoder to use `methodTrampolineLookup`
for the call target whose method object returns false for the `isHelper()` query.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>