Encode only valid XML characters in escapeXMLString()
The function came under focus due to eclipse-openj9/openj9#17978
because the character sequence "&#x1;" encodes a character that
is not legal in XML version 1.0.

The function now only uses an escape sequence for legal characters
and replaces illegal characters with the Unicode replacement U+FFFD.
Also the intermediate buffer is no longer needed.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>