Simplify SupportsConst(I|L)Div to Supports(I|L)MulHigh
This is commit is motivated by openj9#17861.
A check whether the underlying target supports 64 bit
multiplication returning the high word is needed
in order to intrinsify java.lang.Math.multiplyHigh

As a side effect, this has simplified the
SupportsConst(I|L)Div checks as they are transitive
to the newly added checks

Note this commit is a duplicate of PR#7082 due to a
synchronized merge dependency. This commit differs
in that getSupportsLowering(I|L)Div is retained
to avoid the same issue for the downstream dependency.

When the downstream dependency is eliminated, a future
change will remove getSupportsLowering(I|L)Div

Signed-off-by: James You <james.you@protonmail.com>