Stop special-casing unevaluated aconst in Power and ARM linkage
Instead, simply evaluate it like we do for other arguments.

These were the only two places where a TR_RamMethodSequence relocation
could be generated for a method other than the outermost method, which
is not expected. All remaining uses of TR_RamMethodSequence are for
compiledMethodSymbol, which always represents the outermost method.