Fix bug related to CodeCache utilization reporting
PR https://github.com/eclipse/omr/pull/7129 moved the code
that writes a TR::CodeCache pointer into the beginning of
the code cache segment towards the end of the
OMR::CodeCache::initialize routine. This makes sense
because we want to 'advertise' to the world the newly created
TR::CodeCache structure only after this structure has been
fully initialized.
However, OMR::CodeCache::initialize() routine contains an early
return that is taken when the JIT does not need trampolines. On
this path we no longer write the TR::CodeCache pointer into the
beginning of the code cache segment and the VM is unable to
retrieve code cache utilization statistics.
This commit avoid the early return mentioned above.

Signed-off-by: Marius <mpirvu@ca.ibm.com>