Option to reduce counts when class is not in SCC
This commit introduces a new option whose effect is to
reduce initial invocation counts for methods whose
class is loaded during startup but not present in the
Shared Class Cache.
There are actually two opposite options:

-Xjit:useLowerCountsForNonSCCMethodsDuringStartup
-Xjit:dontUseLowerCountsForNonSCCMethodsDuringStartup

The options are mostly for the benefit of the OpenJ9 project.

The option bit `TR_UseLowerCountsForNonSCCMethodsDuringStartup`
is set by default to 1 (`true`), but if aggressive throughput is
desired, the option is reset to 0 (`false`). The user can then
use the command line options mentioned above to override
the default chosen value.
Note that if the user specifies a count value on the command line,
(e.g. -Xjit:count=100), `TR_UseLowerCountsForNonSCCMethodsDuringStartup`
is always set to 0.

Signed-off-by: Marius <mpirvu@ca.ibm.com>