Use small page size/flags for Sparse Virtual Memory
Currently Sparse Virtual Memory uses page size Object Heap is allocated
with. We want to use small pages for Sparse Heap always regardless. Also
Sparse Virtual Memory uses page flags "requested" for Object Heap. This
is the bug. Page flags should be paired with the page size.

Please note, that current MM_SparseVirtualMemory implementation can use
double mapping instead of off-heap. This is an experimental feature
never used in the code and not planned to be used. This feature uses
hidden assumption that memory page size/flags should be the same for
Object and Sparse heaps. This PR can break it potentially in the case
memory page settings in the Object Heap and Sparse heap are different.
We are going to be aware about this if we need another round of
experimental testing with double mapping support.

Signed-off-by: Dmitri Pivkine <Dmitri_Pivkine@ca.ibm.com>