Fix collection of object files for omr_static_lib
The behavior of GNU make seems to have changed between versions
4.2.1 and 4.3.

The latter appears to prioritize MAKEFLAGS= over (at least some)
command-line options, with the effect being that --no-print-directory
is ignored and included in the output of show-objects.

If invoked as part of a larger build where the make job server is
enabled, there are also warnings that make unwanted contributions
to the output of show-objects.

This fixes that, and removes the useless "echo" rule for the already
existing objects.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>