Return another error if only one data point has been recorded
This is to determine if a call to omrsysinfo_get_CPU_load has only one
data point recorded.

Also fixed a mistake in the description saying that the first two
invocations return OMRPORT_ERROR_OPFAILED and fix the time unit in the
description to "ms".

Related: https://github.com/eclipse-openj9/openj9/pull/18451

Signed-off-by: Gengchen Tuo <gengchen.tuo@ibm.com>